wip: skill point and passive point system updates
