+++ Content/GUI/NewHotbar.cs
@@ -1,11 +1,12 @@
 ﻿using Microsoft.Xna.Framework.Graphics;
 using PathOfTerraria.Core.Loaders.UILoading;
 using PathOfTerraria.Core.Systems;
-using PathOfTerraria.Core.Systems.SkillSystem;
 using ReLogic.Graphics;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using PathOfTerraria.Core.Mechanics;
+using PathOfTerraria.Core.Systems.ModPlayers;
 using Terraria.GameContent;
 using Terraria.GameContent.UI;
 using Terraria.GameInput;
