+++ Core/Systems/Affixes/ItemTypes/WeaponAffixes/PassiveAffixes.cs
@@ -5,6 +5,7 @@ public class PassiveAffixes
 	internal class RapidItemAffix : ItemAffix
 	{
 		public override ItemType PossibleTypes => ItemType.Melee;
+
 		public override void ApplyAffix(EntityModifier modifier, PoTItem item)
 		{
 			modifier.Attackspeed += (Value * 5 + Value * item.ItemLevel / 10f) / 100f;
@@ -14,6 +15,7 @@ public class PassiveAffixes
 	internal class SharpItemAffix : ItemAffix
 	{
 		public override ItemType PossibleTypes => ItemType.Melee;
+
 		public override void ApplyAffix(EntityModifier modifier, PoTItem item)
 		{
 			modifier.Damage += (Value * 5 + Value * item.ItemLevel / 20) / 100f;
@@ -23,6 +25,7 @@ public class PassiveAffixes
 	internal class ChargedItemAffix : ItemAffix
 	{
 		public override ItemType PossibleTypes => ItemType.Melee;
+
 		public override void ApplyAffix(EntityModifier modifier, PoTItem item)
 		{
 			modifier.Damage *= 1 + (Value * 5 + Value * item.ItemLevel / 20) / 100f;
