+++ Core/Systems/ModPlayers/SkillPlayer.cs
@@ -1,0 +1,134 @@
+﻿using System.Collections.Generic;
+using System.Linq;
+using Microsoft.Xna.Framework.Input;
+using PathOfTerraria.Core.Mechanics;
+using Terraria.GameInput;
+using Terraria.Localization;
+using Terraria.ModLoader.Core;
+using Terraria.ModLoader.IO;
+
+namespace PathOfTerraria.Core.Systems.ModPlayers;
+
+internal class SkillPlayer : ModPlayer
+{
+	public static ModKeybind Skill1Keybind;
+	public static ModKeybind Skill2Keybind;
+	public static ModKeybind Skill3Keybind;
+
+	public Skill[] Skills = new Skill[3];
+	
+	public int Points;
+
+	public override void Load()
+	{
+		if (Main.dedServ)
+		{
+			return;
+		}
+
+		Skill1Keybind = KeybindLoader.RegisterKeybind(Mod, "UseSkill1", Keys.D3);
+		Skill2Keybind = KeybindLoader.RegisterKeybind(Mod, "UseSkill2", Keys.D4);
+		Skill3Keybind = KeybindLoader.RegisterKeybind(Mod, "UseSkill3", Keys.D5);
+
+#if DEBUG
+		// This code only runs in debug as we don't need to do this in production; keys would already be generated by this point.
+		// This code autogenerates the keys for both the name and description of a skill.
+
+		IEnumerable<Type> skills = AssemblyManager.GetLoadableTypes(Mod.Code).Where(x => typeof(Skill).IsAssignableFrom(x) && !x.IsAbstract);
+
+		foreach (Type skillType in skills)
+		{
+			var skill = Skill.GetAndPrepareSkill(skillType);
+
+			Language.GetOrRegister("Mods.PathOfTerraria.Skills." + skill.Name + ".Name", () => "");
+			Language.GetOrRegister("Mods.PathOfTerraria.Skills." + skill.Name + ".Description", () => "");
+		}
+#endif
+	}
+
+	public override void ProcessTriggers(TriggersSet triggersSet)
+	{
+		if (Skill1Keybind.JustPressed && Skills[0] != null)
+		{
+			if (Skills[0].Timer == 0)
+			{
+				Skills[0]?.UseSkill(Main.LocalPlayer);
+			}
+		}
+
+		if (Skill2Keybind.JustPressed && Skills[1] != null)
+		{
+			if (Skills[1].Timer == 0)
+			{
+				Skills[1]?.UseSkill(Main.LocalPlayer);
+			}
+		}
+
+		if (Skill3Keybind.JustPressed && Skills[2] != null)
+		{
+			if (Skills[2].Timer == 0)
+			{
+				Skills[2]?.UseSkill(Main.LocalPlayer);
+			}
+		}
+	}
+
+	public override void ResetEffects()
+	{
+		if (Skills == null || Skills.Length == 0)
+		{
+			return;
+		}
+
+		foreach (Skill skill in Skills)
+		{
+			if (skill is { Timer: > 0 })
+			{
+				skill.Timer--;
+			}
+		}
+	}
+
+	public override void SaveData(TagCompound tag)
+	{
+		for (int i = 0; i < Skills.Length; i++)
+		{
+			Skill skill = Skills[i];
+
+			if (skill is null)
+			{
+				return;
+			}
+
+			TagCompound skillTag = new()
+			{
+				{ "type", skill.GetType().AssemblyQualifiedName }
+			};
+
+			skill.SaveData(skillTag);
+			tag.Add("skill" + i, skillTag);
+		}
+		
+		tag["points"] = Points;
+	}
+
+	public override void LoadData(TagCompound tag)
+	{
+		Points = tag.GetInt("points");
+		for (int i = 0; i < Skills.Length; ++i)
+		{
+			if (!tag.ContainsKey("skill" + i))
+			{
+				return;
+			}
+
+			TagCompound data = tag.GetCompound("skill" + i);
+			string type = data.GetString("type");
+
+			var skill = Skill.GetAndPrepareSkill(Type.GetType(type));
+			skill.LoadData(data);
+
+			Skills[i] = skill;
+		}
+	}
+}
